got thing to compile and also demo screen kinda cant figure out how to test on phone tho
